1.modify about AutoLock; 2.fix bug about return; 3.fix compiler;

+++ Sources/Elastos/LibCore/src/elastos/math/CBigDecimal.cpp
@@ -463,7 +463,7 @@ ECode CBigDecimal::ValueOf(
     VALIDATE_NOT_NULL(result);
 
     if ((unscaledVal >= 0) && (unscaledVal < BI_SCALED_BY_ZERO_LENGTH)) {
-        *result = (*BI_SCALED_BY_ZERO)[unscaledVal];
+        *result = (*BI_SCALED_BY_ZERO)[(Int32)unscaledVal];
         REFCOUNT_ADD(*result);
         return NOERROR;
     }
@@ -2657,7 +2657,7 @@ ECode CBigDecimal::DoubleValue(
 
     if ((powerOfTwo < -1074) || (sign == 0)) {
         // Cases which 'this' is very small
-        *result = (sign * 0.0d);
+        *result = (sign * 0.0l);
         return NOERROR;
     }
     else if (powerOfTwo > 1025) {
@@ -2762,7 +2762,7 @@ ECode CBigDecimal::DoubleValue(
     else if (exponent <= 0) {// (exponent - bias <= -1023)
         // Denormalized numbers (having exponent == 0)
         if (exponent < -53) {// exponent - bias < -1076
-            *result = (sign * 0.0d);
+            *result = (sign * 0.0l);
             return NOERROR;
         }
